``/GET/{taskId}`` route secured

+++ Controllers/TasksController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using ToDoApi.Models;
 using ToDoApi.Services;
+
 namespace ToDoApi.Controllers
 {
     [Authorize]
@@ -34,6 +35,11 @@ namespace ToDoApi.Controllers
         {
             var toDoItem = await _tasksService.GetAsync(id, Util.User.GetUserId(HttpContext));
 
+            if (toDoItem is null)
+            {
+                return NotFound();
+            }
+
             return ItemToDto(toDoItem);
         }
 
@@ -47,10 +53,14 @@ namespace ToDoApi.Controllers
             }
 
             var toDoItem = await _tasksService.GetAsync(id, Util.User.GetUserId(HttpContext));
+            if (toDoItem is null)
+            {
+                return NotFound();
+            }
 
-            if (toDoItem == null) return NoContent();
             toDoItem.Name = taskItemDto.Name;
             toDoItem.IsComplete = taskItemDto.IsComplete;
+
             await _tasksService.UpdateAsync(id, toDoItem);
 
             return NoContent();
@@ -79,12 +89,16 @@ namespace ToDoApi.Controllers
         public async Task<IActionResult> DeleteToDoItem(string id)
         {
             var toDoItem = await _tasksService.GetAsync(id);
+            if (toDoItem is null)
+            {
+                return NotFound();
+            }
 
             await _tasksService.DeleteAsync(id);
             return NoContent();
         }
 
-        private async Task<bool> ToDoItemExistsAsync(string id)
+        private async Task<bool> TaskExistsAsync(string id)
         {
             var tasks = await _tasksService.GetAsync(id);
             return tasks.Any(x => x.Id == id);
